Max_raw_code/requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
pydantic==2.8.2